Fixed distance. Debugged axis, added new filter type.

+++ 2017 Robot/src/Vector/Vector2D.java
@@ -99,7 +99,7 @@ public class Vector2D {
 		this.x += x;
 		this.y += y;
 	}
-	public void addVectorWithTime(Vector2D otherVector, long t) {
+	public void addVectorWithTime(Vector2D otherVector, double t) {
 		this.x += otherVector.getX() * t;
 		this.y += otherVector.getY() * t;
 	}
